requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.14
fastapi==0.62.0
funcy==1.15
future==0.18.2
gensim==3.8.3
gunicorn==19.9.0
h11==0.11.0
httptools==0.1.1
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
nltk==3.5
numexpr==2.7.1
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pydantic==1.7.3
pyLDAvis==2.1.2
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
scipy==1.5.4
six==1.15.0
smart-open==4.0.1
starlette==0.13.6
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.1
wincertstore==0.2
zipp==3.4.0